+++ examples/simulations/hanoi3d.py
@@ -3,7 +3,7 @@
 # https://github.com/gjbex/training-material/blob/master/Misc/Notebooks/hanoi.ipynb
 # Creative Commons Zero v1.0 Universal licence
 
-from vedo import Plotter, Cylinder, Box, ProgressBar
+from vedo import Plotter, Cylinder, Box
 from copy import deepcopy
 
 
@@ -65,9 +65,7 @@ cam = dict(
 )
 plt.show(camera=cam)
 
-pb = ProgressBar(0, len(tower_states), 1, c="y")
-for t in pb.range():
-    pb.print()
+for t in range(len(tower_states)):
     state = tower_states[t]
     for tower_nr in range(3):
         for i, disk in enumerate(state[tower_nr]):