+++ docs/changes.md
@@ -5,6 +5,8 @@
 - method `tomesh()` changed to `generate_mesh()`
 - fix for `mark_boundaries()`
 - use vtkStaticCellLocator instead of vtkCellLocator
+- improved `pointcloud.density()` function
+- add `utils.progress_bar()`
 
 -------------------------
 ## Examples